pyproject.toml
[tool.poetry]
name = "emmett"
version = "0.1.0"
description = "A mistake."
authors = ["Slavfox <slavfoxman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.5.1"
markovify = "^0.8.3"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
known_third_party = ["emmett"]
force_to_top = ["typing"]
